refactor : move Generator to factory package
